Update using bool indicator with global scope
I'm suggesting these changes as I noticed a minor issue with this bit of code. When I input "the birds fly why fish swim" it showed the result "Not okay" \n "Ok", instead of just "Not Okay" 

It seemed like this was due to the bool indicator for an error being located within the while loop. I have changed sentence() and conjunction() from bool to void, as they no longer need to return a value to main(). The program now remembers if there has been a mistake, even when going back through sentence(), and delivers one answer after the prompt "." rather than multiple answers for one sentence.